Create table creator class.
